+++ app/dance/spinners/heart.go
@@ -7,18 +7,18 @@ import (
 )
 
 type HeartMover struct {
-	start int64
+	start float64
 }
 
 func NewHeartMover() *HeartMover {
 	return &HeartMover{}
 }
 
-func (c *HeartMover) Init(start, end int64) {
+func (c *HeartMover) Init(start, end float64) {
 	c.start = start
 }
 
-func (c *HeartMover) GetPositionAt(time int64) vector.Vector2f {
+func (c *HeartMover) GetPositionAt(time float64) vector.Vector2f {
 	rad := rpms * float32(time-c.start) * 2 * math32.Pi
 	x := math32.Pow(math32.Sin(rad), 3)
 	y := (13*math32.Cos(rad) - 5*math32.Cos(2*rad) - 2*math32.Cos(3*rad) - math32.Cos(4*rad)) / 16
